Prueba

+++ ServerTiempo.py
@@ -148,8 +148,8 @@ def createClientThread(connection, c):
         #print ("nuevas Horas: "+ str(nuevasHoras))
         tiempo = []
         while i in range (4):
-            horaaux = nuevasHoras[0]//60
-            minutosaux = nuevasHoras[0] - (horaaux*60)
+            horaaux = int(nuevasHoras[0]//60)
+            minutosaux = int(nuevasHoras[0] - (horaaux*60))
             tiempo.append(str(horaaux)+":"+str(minutosaux)+":"+"00")
         print(tiempo)
     c.close()
